+++ src/tkenv/tkenv.cc
@@ -547,7 +547,7 @@ bool Tkenv::doRunSimulation()
 
         // display update, even if we already did one before the animations
         if (frequent_updates || ((getSimulation()->getEventNumber()&0x0f) == 0 && elapsed(opt->updateFreqFast, last_update))) {
-            if (!frequent_updates || (speedometer.getMillisSinceIntervalStart() > opt->updateFreqFast && speedometer.getNumEventsSinceIntervalStart() >= 3)) // do not start new interval at every event
+            if (!frequent_updates || ((int)speedometer.getMillisSinceIntervalStart() > opt->updateFreqFast && speedometer.getNumEventsSinceIntervalStart() >= 3)) // do not start new interval at every event
                 speedometer.beginNewInterval();  // should precede updateStatusDisplay()
             callRefreshDisplay();
             refreshInspectors();
